Move operator parsing to internal folder
