+++ src/internal/expr-interp-rotate.c
@@ -1,5 +1,5 @@
 #include <rlang.h>
-
+#include "parse.h"
 #include "expr-interp.h"
 #include "expr-interp-rotate.h"
