+++ src/rlang/rlang.c
@@ -112,15 +112,6 @@ SEXP r_init_library() {
   r_quo_get_env = (sexp* (*)(sexp*)) r_peek_c_callable("rlang", "rlang_quo_get_env");
   r_quo_set_env = (sexp* (*)(sexp*, sexp*)) r_peek_c_callable("rlang", "rlang_quo_set_env");
 
-  /* parse.c - r_ops_precedence[] */
-  RLANG_ASSERT((sizeof(r_ops_precedence) / sizeof(struct r_op_precedence)) == R_OP_MAX);
-
-  for (int i = R_OP_NONE + 1; i < R_OP_MAX; ++i) {
-    if (r_ops_precedence[i].power == 0) {
-      Rf_error("Internal error: `r_ops_precedence` is not fully initialised");
-    }
-  }
-
   // Return a SEXP so the init function can be called from R
   return r_null;
 }
